requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
jsonpatch==1.32
jsonpointer==2.3
numpy==1.22.3
Pillow==9.1.0
pycodestyle==2.8.0
pytorch-fid==0.2.1
pyzmq==22.3.0
requests==2.27.1
scipy==1.8.0
six==1.16.0
toml==0.10.2
tornado==6.1
typing-extensions==4.2.0
urllib3==1.26.9
visdom==0.1.8.9
websocket-client==1.3.2
